LayoutTests/imported/w3c/web-platform-tests/page-visibility/test_read_only-expected.txt

Description

This test validates that the Page Visibility attributes are read only.


PASS document.hidden is defined and not null. 
PASS document.visibilityState is defined and not null. 
PASS document.hidden is read only. 
PASS document.visibilityState is read only. 
